#7: Implemented setting of page size on sections and whole document

+++ src/main/java/org/wordinator/xml2docx/generator/Measurement.java
@@ -95,9 +95,9 @@ public class Measurement {
 	 * @return Number of twips
 	 * @throws MeasurementException Bad measurement value 
 	 */
-	public static int toTwips(String measurementValue, int dotsPerInch) throws MeasurementException {
+	public static long toTwips(String measurementValue, int dotsPerInch) throws MeasurementException {
 		double points = toPoints(measurementValue, dotsPerInch);
-		return (int) points * 20;
+		return (long) points * 20;
 	}
 
 }
